+++ src/controllers/clothes/controllersPostClothes.js
@@ -1,11 +1,12 @@
 const mongoose = require('mongoose');
-const clothes = require('../models/clothes')
+const clothes = require('../../models/clothes')
 
 const postTheClothes = async ({title,description,image,price,stock,color,gender,category}) => {
 
   try {
 
     if(!title || !description || !image || !price || !stock || !color || !gender || !category) throw new Error("Faltan completar los datos correctamente")
+    
     else{
       const resultNewClothes = new clothes({title,description,image,price,stock,color,gender,category})
       await resultNewClothes.save()
